unsealed/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unsealed"
version = "0.4.1"
description = "Program to decode and encode various game files from online game: Seal Online."
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "feryandi", email = "feryandi.n@gmail.com" }]
classifiers = ["Programming Language :: Python :: 3"]

# Project Dependencies
dependencies = [
  "autopep8==2.3.2",
  "scipy==1.17.0",
  "numpy==2.4.1",
  "setuptools==80.10.1",
  "pillow==12.1.0",
  "pygltflib==1.16.5",
  "Sphinx==9.1.0",
  "ruff==0.14.14",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black", "isort", "mypy"]

[project.scripts]
unsealed = "unsealed.__main__:run"

[project.urls]
"Homepage" = "https://github.com/feryandi/unsealed"
"Bug Tracker" = "https://github.com/feryandi/unsealed/issues"

[tool.setuptools]
package-dir = { "" = "src" }
packages = { find = { where = ["src"] } }
